Move schema settings from settings to sidebar
