+++ frontend/src/locales/pl.json
@@ -211,7 +211,7 @@
       "eventErrorSchemaExists": "Schema already exists"
     },
     "schemaSettings": {
-      "title": "Schema Settings",
+      "title": "Schemas",
       "buttonCreateSchema": "Create Schema",
       "buttonImportSchema": "Import Schema"
     },
