+++ frontend/src/views/SchemaSettings.vue
